Refactor, and add list of all rounds in game.
Refactor out common card template stuff into its own template file.
Refactor out returning an error from an endpoint into a common function.
Add general logging and configuration for it.
Define an interface for handlers so it's easier to add more without maintaining a centralized list.
Add a list of all rounds in a game, and link to it from a round. It links back to the full round details as well.